+++ CHANGELOG.md
@@ -7,7 +7,8 @@ We use the following versioning scheme `YEARMMDD` so version `20190308` would ha
 ## 20190329
 
 - Remove caching functionality from user service to prevent incorrect data.
-
+- Add javadoc site links
+- Refactor UserActivity#principal from String to User.
 
 ## 20190308